Set the API key, make various pages, new directives
Now the app will be able to get your location from your IP and then
fetch the forecast for you.  This requires an API key and I don't want
to commit this to the repository, so a new screen was made to prompt the
user for their personalized API key, which is then saved in local
storage.